[network-mgr] Adds configuration for qemu
Due to the way that topological paths currently work, when running
under qemu the emulated topological path needs to also be configured
so that network manager can configure the emulated interface correctly.

This change introduces a `default_interface` that can be used to
configure these unconfigured devices.

Also, this change narrows the tests that the router QEMU buildbot runs
because, for example, are not including the bluetooth or lowpan stacks.

Bug: 50780
Change-Id: If2f0942462a0f45b852f2b0a225abae376fccc20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383881
Commit-Queue: Chris Gibson <cgibson@google.com>
Reviewed-by: Nicolas Sylvain <nsylvain@google.com>
Reviewed-by: Chris Gibson <cgibson@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Joshua Liebow-Feeser <joshlf@google.com>