[archivist] Report "ERROR:" klogs as Severity::Error
In order to get a better read as to the number of error logs
produced by the system, we need to count klogs as well. Prior to
this CL, all klogs were reported as Severity::Info. This CL is a short
term workaround, until we have structured klogs, to look for the string
"ERROR:" and "WARNING:" to map to Severity::Error, and Severity::Warn
respectively.

Bug: 51179
Change-Id: If63e960eb38cd0c30dda1280c98d8f58135d5626
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384900
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>