[runtests] Update output
This change tidies up redundant output in the case of a single test
being run. Moreover, it adds documented signatures to the output that
characterize a success or failure. In a future CL, this will be used to
determine just that for tests run over serial.

Bug: 41930
Change-Id: I47e4b7035082c35329cd2b52a6abd05da1d63828
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383871
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>