[modular] Re-land "[sessionmgr] Replace ledger with assorted std::map"
A bug in the previous iteration of this change would notify clients of
StoryProviderWatcher of a story being deleted even in the case that the
story was already deleted.

It turned out one of our clients was relying on this behavior, but there
was no test coverage :\. This CL is landing after a change which adds
the required test coverage.

Bug: 49677
Test: fx test session_shell_test; Also manually verified the test fails without the fix.
Change-Id: I47875cd9d7e3e9a32f1fe449e78873509a592634
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385000
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>