[build] Unsplit uboringssl from boringssl
For historical reasons, there were three separate build targets
for boringssl provided in-tree:

* //third_party/boringssl, which is where we maintain a copy of upstream
  boringssl with some build changes to support Fuchsia
* //zircon/third_party/ulib/boringssl, which reexposed the first to
  the zircon build system
* //zircon/third_party/ulib/uboringssl, which once contained a tightly
  restricted subset of source from boringssl for use in the kernel.

We've mostly deduped these, to the point that the second item in that
list is just a symlink farm of the source from the first and BUILD.gn
rules.   The third is just a GN target forwarding to the second, and
serves no particular purpose.

We can eliminate the third target by updating all references that point
to uboringssl to point to the second.

We won't be able to eliminate the second until build unification is
completed, including for the zircon kernel, or the kernel stops making
use of the zircon-exported library, but we could in the meantime migrate
code in the Fuchsia (not zircon) build from the second option to the
first option.  We leave that to future work.

Fixed: 42701
Change-Id: Ia325202a55836a2aefe6d188aeae9fbead78168c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385197
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Testability-Review: Bryan Henry <bryanhenry@google.com>