[logger][sshutil] Add method to extract a logger from a context
This adds a function `LoggerFromContext`. This is particularly helpful
for `sshutil.Conn`, so we can reuse the logger, but not have to worry
about the context we use to control the initial connection to the server
to cancel the keepalive thread early.

Change-Id: If7fe6923f8b77dd82fc3b269ffcdbccb27fba125
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386373
Reviewed-by: Jake Ehrlich <jakehehrlich@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>