[camera][scenic] Conditionally set sysmem buffer names.
This change makes the camera stack and scenic ImagePipe2 less aggressive
in setting the debug name of objects. Instead of unconditionally setting
the names of negotiated sysmem buffers, they use the MaybeSetObjectName
function to set the name only if it is still the sysmem-default name.
Additionally, this change alters where scenic sets the name of buffers
passed to ImagePipe2, setting them only when a specific buffer is added
via AddImage instead of when the collection is initially added.

Test: Used memgraph tool to verify expected debug object names.
      Extended DeviceTest.RequestStreamFromController to check vmo names
        provided by camera.
Change-Id: If66cf64b28b8769af30a280397bf023a64a0277f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385976
Commit-Queue: Matt Sandy <msandy@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>