[components] Use the Rust crate paste to simplify events macro
Instead of explicitly naming the payload and error structs in the
create_event! macro, we use the paste::item! macro to concatenate
Payload, and Error to the end of the event name to name the payload
and error objects respectively.

Change-Id: Ic4aa260253bd1f40f5a2a5bef2e707669b1c13db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386337
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Fady Samuel <fsamuel@google.com>