[GN SDK][CQ Tests] Run GN SDK script tests in CQ
This adds the tests in //scripts to the test bundle and the
buildbot:config.

These are unit tests for the f* scripts which are shared with
teams that are building components in GN based projects.

Multiply: gn_sdk_script_tests

Change-Id: I63794c5423848b791bd74326d8a2cc260c6de8fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/358229
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Matt Carroll <matthewcarroll@google.com>
Testability-Review: James Tucker <raggi@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>