[gidl] Test recursive unions in GIDL
This CL adds GIDL conformance tests for unions that recursively contain
themselves (1) in a vector and (2) via mutual recursion with a struct.

The tests are disabled in Go because using these types causes a stack
overflow (fxb/38191), and in Dart because the Dart analyzer complains
about self-references (fxb/7644). In the case of Dart, we also have to
hide the FIDL declarations using BindingsDenylist, because the analyzer
problem happens when compiling the bindings.

Test: run all GIDL tests
Fixed: 36424
Change-Id: I6d7c738b5c63b11947692979549f865ea9d3f826
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386198
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Testability-Review: Felix Zhu <fcz@google.com>