[SetUI][core] Remove startup result receiver.
A result receiver was previously handed back with the
environment on spawn in order to allow spawn to be
synchronous, but a caller (mostly tests) to monitor
the end result.

Recent changes have modified this code path to
allow for spawn be awaited on by executing on a
single thread. Doing so allows for waiting on the
environment to bring up before returning.

This change makes all environment spawns
awaited on and removes the callback receiver.

Fixed: 50378
Test: fx test -o setui_service_tests

Change-Id: I2b92ea62c45e91a7d649994cc20e14adff7994ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381434
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>