[sdk] Properly handle the deps attribute of sdk_source_set.
Previously this attribute was used to attach dependencies that would not
be reflected in SDKs. This means that public_deps had to be used for these
even for dependencies only used in the implementation of the source set.

Instead, deps are now added to SDK deps, and a new non_sdk_deps should be
used to attached non-SDK deps (e.g. codegen actions).

Change-Id: I05e1f32336a45025f5fc4abd124c4aedc38d711e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386739
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: P.Y. Laligand <pylaligand@google.com>