[run_test_component] Add config to allow Error logs
Adding config to allow error logs for certain test urls.

The feature to fail on ERROR is not active yet and this is first step in
implementing that feature.

The config is read from base package, so developers need to do fx update
when adding a test url to the config.

BUG=49735

TEST=added unit tests

Change-Id: Ia5c8dbea7d2e68d3b3b9b2ef09d10e736bfd0742
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386433
Commit-Queue: Ankur Mittal <anmittal@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Shai Barack <shayba@google.com>