[kernel][wait] Encapsulate the data structure inside wait_queue_t
We'd like to eventually change this data structure. By giving it
names, we are taking a step to separating the data structure (i.e.,
how the collection of threads is laid out, adding to or removing from
the collection, finding a thread in the collection) from the wait
queue operations (waking a thread, etc.).

Change-Id: I9468d9b91bd405bd6c46522bd76b77300da2df2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385232
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: George Kulakowski <kulakowski@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>