[camera][examples] Delete isp_stream_provider and isp-tester fidl
This is a reland of 383201. The original CL was reverted due to relying
on another CL (383198) which caused a flake. The flaky CL has been fixed
and this one is being landed on top of it as-is.

Tested: n/a

Change-Id: I95be64bc3470923cc8b5b84c5971178424119d10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385129
Commit-Queue: Nic Zoghb <nzo@google.com>
Reviewed-by: Brijen Raval <braval@google.com>
Testability-Review: Brijen Raval <braval@google.com>
API-Review: Ruchira Ravoori <ravoorir@google.com>