[paver] fix sherlock bootloader read and size
We had previously fixed SherlockBootloaderPartitionClient to properly
skip the block0 metadata when writing, but reads were still starting at
block0, with the effect that we would pave the bootloader every time
since the check for content equality always showed a mismatch.

With this CL, the paver properly skips writing the bootloader if it's
the same as what's already in the partition.

This also fixes the reported partition size to omit the first block,
since it's not really writable.

Test: `fx pave` and `fx ota` correctly skip matching bootloader image
Bug: 50999
Change-Id: Icbec4dc15bbc1a5ceb9839f66105bcbf811f4394
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384914
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Pursell <dpursell@google.com>