[swd-e2e-tests] Don't use `amberctl ... -verbose`
We added `amberctl ... -verbose` to help debug the random
disconnections, on the off chance it was `amberctl` doing something
weird. We eventually tracked this down to ssh issues in 47145.
Unfortunately, this conflicts with testing from updating from an old
Fuchsia to a newer one. We didn't actually get much value from this,
especially now that we've added more instrumentation to the e2e tests
and the package resolver.

Test: Verified OTA tests works without the `-verbose` flag.
Bug: 50652
Change-Id: I7bc194e34c7067d2e36a6ace3e35b763d9ac5b16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387057
Testability-Review: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>