[region-alloc] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow regions to exist in two trees at once.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: I38784646660d5f03fad47f4f31607e3aecc9f343
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386815
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>