[vmo-mapper] Remove unused vmar_info from ResizeableVmoMapper::Grow()
The |vmar_info| was a remnant from old code that used the base of the
VMAR to determine the location for extending the mapping. It is unused
now. Get rid of the unnecessary zx_object_get_info() syscall.

Test: No change in behavior
Change-Id: Ic59eb2b844875e513e4aca48cc47e52d2488629c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386468
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>