[rust] Add hypen to extra-filename
Based on how cargo does this, and how the -o filename is set, and how
the dep-info is set, I am pretty certain that we want filenames to be
like: libfoobar-123deadbeef.rlib instead of libfoobar123deadbeef.rlib

It looks like gnaw missed a hyphen, so this adds it back, fixes the
tests and updates the third_party build file. This wouldn't be visible
in practice because we specify the output filename explicitly, but it
shows up if you try to reuse the generated command for other things.

Change-Id: Ib6c1c2403e681411c60460734f815c633307b4be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385253
Commit-Queue: Benjamin Brittain <bwb@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Benjamin Brittain <bwb@google.com>