[components] Ignore event_dispatch_order_test
It's a fundamentally flaky test that exposes a race in the Discovered
event. I don't believe the Discovered event really ever makes sense as a
blocking event. It also can easily be missed when subscribing to events.

This CL disables the test for the time being.

Discovered makes sense as an internal event. Externally, we might want
to consider a synthetic version of Discovered.

Bug: 49314
Change-Id: Ibc69942baa7736709cfa0971c19cf3f23e8c68c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387058
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: Fady Samuel <fsamuel@google.com>