[build] Enable linker GC flexibly based on `optimize` setting
This changes config("default_linker_gc") to dispatch to a config()
selected directly by the `optimize` build argument, making it an exact
parallel to config("default_optimize").  As with default_optimize,
this scheme is uniform between Zircon and Fuchsia builds.

In Zircon, the kernel's special case to force linker GC is
accomplished in a cleaner fashion, and is now limited to only x86
where it is needed.  The only configuration that should be changed by
this refactor is that the `optimize="none"` (formerly `opt_level=-1`)
arm64 build's kernel will no longer use linker GC.

Bug: 51507
Change-Id: I85d2de7706abdde7a4bb2a29f6f16216b40696aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386457
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>