[pci] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
maintain a tree of devices.  Also removes some uses of "mutable" on
locks that was not needed, and renamed the "device_list_" member of
bus to just "devices_", as this was not actually a list of devices but
a tree.

This refactor removes a multiple uses of custom container traits from
the system which helps to prepare for an up-and-coming refactor of the
fbl:: container template arguments which will require special handling
of users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: I2f70d0a0ecf1cb0f87666d62f309848df7b9ffb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386825
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>