[build][breakpad] Delete breakpad archive creation
This is no longer needed(!) and neither is the underlying tool that
creates it (dump_breakpad_symbols). In this change, we wave goodbye to
ugly code and hackery (the GN integration), and thank it for its
service. A follow-up will delete ids.txt altogether.

Bug: 49782
Change-Id: Ia73713bcaadaf87b244eb6790e6fa194d6cbc42f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386746
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Jake Ehrlich <jakehehrlich@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Petr Hosek <phosek@google.com>