[fuchsia.devicesettings] remove topaz visibility from fuchsia.devicesettings
This CL removes fuchsia.devicesetting SDK visibility from topaz.

Deprecated settings code is being migrated to experiences so it can be removed
from the topaz repo (and eventually removed from experiences once new settings
feature parity is met).

Bug: 45550

This is 4 of 4 CLs that need to be submitted in order:
1. [fxr/381099] add SDK visibility to experiences/settings
2. [fxr/380992] copy lib.settings into experiences
3. [fxr/381113] delete lib.settings from topaz
4. => [fxr/381106] remove SDK visibility from topaz

The set of CLs has been vaildated with:

        fx make-integration-patch

Change-Id: I1de7ccf310c556a432162a27e99525d55fb8442f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381106
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Jason Campbell <jasoncampbell@google.com>