[ermine] Remove login_shell.
Since Ermine migrated to SfW this is now unused and modular is in the
process of removing the base shell interface that it relied upon. Some
day in the not too distant future it will rise again, like a phoenix
from the ashes, in an even more glorious session framework form.

Bug: 48782
Change-Id: I2464023cab728a9b84bd3d7c2050e58bc4b47514
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/386716
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Testability-Review: Chase Latta <chaselatta@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>