[dispatcher-pool] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
implement lists used to manage EventSources in their ExecutionDomains.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Change-Id: I68a20fe38a4b76c5b0bb9223bcb372384fffaf83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386814
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>