[kernel][pmt] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
exist simultaneously in multiple lists, and start using the
fbl::ContainableBaseClasses feature instead.

In addition to (hopefully) making the code a bit easier to understand,
this refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Also, it turns out the PMT Dispatchers were holding 3 different DLL node
state in themselves, but using only two.  In addition to the custom
PmtList node, and the QuarantineList node, the class was also deriving
from DoublyLinkedListable which introduced a 3rd (unused) node into
the mix.  This has now been removed, which means that objects should
be at least 2 pointers smaller now.

Tests: `k ut all` and `runtests` on a bringup cavium build
Change-Id: Id31b917294c7582f6b0bf997a158cbc0456b5f86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386749
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>