[blobfs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
track cached blobfs nodes.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: If09d3932558cbabd9067655322aff142892839bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386816
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Christopher Anderson <cja@google.com>