[build] Derive unique test name in test spec
Previously the name field was an arbitrary string passed by the caller.
Because it was not guaranteed to be unique, it was mostly ignored by the
infrastructure, and instead rewritten using the logic in this CL:
if component URL is present, use that.
else, use path.

With this change, we can remove this logic from the infrastructure and
just use the name field as it appears.

This also simplifies all the call-sites of the test_spec GN template.

Bug: 50400
Change-Id: Ieec89d3e1973dfcb755f1001f779fef51ec4128e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387463
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>