[storage] Update Minfs C++ style.
Add a DirentBuffer class to simplify the task of reading from the device
into a buffer large enough for the largest Dirent and casting it. This
also resolves alignment issues with doing this (although there are a lot
of other places doing unaligned access on filesystem structures).

Use the new DirentBuffer in InitializeDirectory and append the results
with memcpy as a second step. This code is not performance-critical and
the new flow ensures aligned access with less pointer computations and
no casting.

Provide default values for DirArgs. Although not a problem, this class
is not performance critical and default values make it easier to use
correctly.

Rename MINFS_DIRENT_SIZE to kMinfsDirentSize.

There should be no behavior change.

Change-Id: Iab14ccae23315cba54923e6469f42252902d1eb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388013
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>