[benchmarks] Remove non-SL4F-based invocations of most performance tests
Catapult now alerts on the SL4F versions the performance tests and not
the versions being removed here.

I'm not quite ready to remove the peridot-x64-perf-dawson_canyon
builder for the reason explained in the comment, but removing most of
the tests from it will save resources and remove the need for people
to maintain the invocations in src/tests/benchmarks/benchmarks.cc.

Bug: 23091
Test: on CQ: peridot-x64-perf-dawson_canyon builder
Change-Id: I759458bb28cb25a1194524e106eee563e3ad195a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387526
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>