[storage] Verify that the tail of blobs are zeroed.
An attacker that is able to manipulate the disk, could change data
beyond the end of the blob. This change adds verification so that this
should not be possible to go undetected.

In theory, this could bring existing corruption that has been
undetected to light.

Bug: 51377
Test: blobfs-unit-test
Change-Id: I45c064f4b10f4c6c18cb55c85f384ee7f98cc72f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387013
Commit-Queue: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>