[fidlcat] add --extra-name option
This option allows fidlcat to start displaying events only when one main
process is started and allows to shutdown fidlcat when there is no more
main process.

This is useful when you want to monitor some services on top of your
application. It avoids fidlcat to display the services' events which are
not related to your application.

It allows fidlcat to shutdown when your application terminates (because
services never terminate, monitoring a service with --remote-name
instead of --extra-name forbids fidlcat to exit).

Change-Id: If0bd3fc2d66ee24ea5229e6e04d38c055dcedea1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385375
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>