[fidl][benchmarks] Benchmarks for non-byte array/vector
This is to detect when we are overfitting optimizations.
For instance, in Go today uint8 vectors are special cased.

The array/vector elems are non-primitive inline values with padding.
This highlights optimizations that only work when there is padding or
specific primitive types.

Change-Id: Id083af51cd1855537f8bd0d111056e6554a3e90d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388135
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>