[swd-e2e-tests] Add monotonic time prefixer
It can be difficult to trace events across the e2e tests, since the
cause could be distributed across the e2e logs, the system logs, and the
serial logs. Since all the device logs are emitted with the device's
monotonic time, this patch extends `device.Client` to query the current
time on initial connection. Then, `device.WithMonotonicTimePrefix` can
be used with `logger` to prepend an estimate of the current device time
to each log message.

Test: This is unfortunately difficult to test without a real device to
test with, so it was confirmed to work by hand.
Fixed: 50266

Change-Id: If5d751341596e7f205aab8f4cb5b53dd47bce623
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386103
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Ben Keller <galbanum@google.com>
Reviewed-by: Victor Chang <vfcc@google.com>