[testrunner] Add validation check for tests.
This will validate that the tests have the required fields. If the
test schema is updated here but not in the recipes, this could cause
testrunner to silently fail to run tests. This check should catch such
failures.

Bug: 51727
Change-Id: Ic656fa07e4dfe8bc6fc403ad3bb36b454aa1254c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387833
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>