[terminal] Port to render from canvas
This is a somewhat brute-force port of Terminal to render from canvas.
The sole purpose of the CL is to eliminate the last use of canvas so it
can be removed from Carnelian and get rid of a lot of technical debt.

Since the foreground and background color can change per cell, our
usual approach of caching the glyph rasters will not work. Instead
glyphs are created as needed. Once Terminal is unpaused, creating a
sophisticated enough glyph cache would be an excellent task.

This implementation also uses pre-clear, rather than preserving the
rasters to erase them on the next frame. This would be another
excellent task for a resurrected Terminal effort.

It is also possible that these glyph caching and optimal redraw
strategies will end up becoming features of Carnelian and Terminal will
get them for free.

Testing; ran all unit tests and manually tested on core and workstation
builds.

Change-Id: I24752c9443ca316b058bc0a9273be458d0635fe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387388
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>