[fidl][benchmarks] Runner for fidl microbenchmarks
This is intended to be used with the chromeperf runners, but is also
useful for running locally and generating data. There is a permission
issue when running the benchmarks (e.g. go_fidl_microbenchmarks) using
this runner on CQ which I am investigating but I wanted to commit this
as-is because it is useful for generating csv data in the meantime by
running locally.

- This currently only supports builder/encode/decode benchmarks
- Produces a standardized output, suitable for tracking in chromeperf
- Can produce csv

Change-Id: Ia40dcb4134123b307b903cd03be0e50e4793b35f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386654
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>