[cobalt] Fix GN targets for forming the cobalt_tests package
The main problem that this addresses is that the
cobalt_fake_timekeeper was in its own package. Discussions with
the Component Framework and Software Delivery teams indicated
that all of the processes involved in an integration test
should be in a single package. This CL moves it into the
cobalt_tests package.

The symptom I was experiencing that this fixes is the following:
   run_test_component cobalt_testapp_no_network
the fake timkeeper was not being started.

The problem turned out to be that the fake timekeeper was not an
available package. I filed fxb/51788 because no error message
is given when this happens.

My fx set command included cobalt_tests and the cobalt_tests package
depended on the cobalt_fake_timekeeper package, so I assumed that
the cobalt_fake_timekeeper package was being included. But it
turns out that a package target is not supposed to depend
on another package target and doing so is a no-op.

I also made some other related fixes to our BUILD.gn files.
I have added code review comments on each line to help a reviewer
understand why I am making each change.

Change-Id: Ice5f15832dc8dbc81596b5a09b5945c048f999a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388258
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>