[archivist] Report "ERROR:" klogs as Severity::Error"
In order to get a better read as to the number of error logs
produced by the system, we need to count klogs as well. Prior to
this CL, all klogs were reported as Severity::Info. This CL is a short
term workaround, until we have structured klogs, to look for the string
"ERROR:" and "WARNING:" to map to Severity::Error, and Severity::Warn
respectively.

This CL addresses a previous issue in the reverted
2ef64fbeae1080e2e15f9698c118d8dce61cea21 patch. Instead of slicing
a string at a fixed byte index, it splits it at a fixed character
index which ensures we handle utf-8 correctly.

Bug: 51179

Change-Id: Ibf679919a049c5ffd8b440344e3e1ad3aec9ea72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387933
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>
Testability-Review: Adam Perry <adamperry@google.com>