[virtualization] Fix virtio-queue index handling
Check that the difference between the ring index and avail index does
not exceed the size of the queue.

In addition, do some more small cleanups in the queue handling logic in
order to be able to remove some of the legacy in-process code in the
future.

Change-Id: I55eaa965b8040035812d9b85a54762b705f79d11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388100
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>