[cobalt][test_app] Utilize the new ListenForInitialized() method.
This fixes a race condition between the Cobalt test_app and the
initialization of the Cobalt service by replacing a sleep
with waiting for the initialization signal.

Bug: 51211
Bug: 51473.

This does not yet fix the Prober. In a follow-up CL I will
fix the Prober by stopping using the FakeTimekeeper with
the prober.

Change-Id: I299662768e4fe020c904d051827ab45a3a765741
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388256
Reviewed-by: Zach Bush <zmbush@google.com>
Testability-Review: Zach Bush <zmbush@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>