[wlan][iwlwifi] port iwl_trans_pcie_log_scd_error()
This function is used later to dump the Tx queue info to debug
in the error cases. Make it standalone so that easier for review.
Most of changes are style.

Bug: 45032
Tested: Compiled only. Will be used in the following CL.

Change-Id: I3a833f482638e57d6c97abdcfa207ab30925acd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385380
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>
Testability-Review: Alex Legg <alexlegg@google.com>