[cobalt] test_app to use real Utc svc when it uses real network
The versions of the Cobalt test_app integration test that use
the real network also need to use the real UTC time service
because they need to know when the real system time has
become accurate.

We also added fuchisa.time.Utc to the allow-list of services
that could be included in the "system-services" of a test.

Bug: 51473
Bug: 49306

Change-Id: Ia6bb98bcd8b0b3ef44a844632d58e04517f63829
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388547
Commit-Queue: Mitch Rudominer <rudominer@google.com>
Reviewed-by: Zach Bush <zmbush@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Testability-Review: Zach Bush <zmbush@google.com>