[ffx][target] Refactor Target to avoid deadlocks.
Refactors some code around the Target struct in order to make it more
difficult to create potential deadlocks. It is now not possible outside
of testing and inside of the target module for there to be more than one
lock acquired simultaneously in a Target struct.

It was not feasible to create a reliable public API test that would have
(previously) created a deadlock without simply acquiring several locks
in step with each other (which is not possible now with this change),
so this only includes tests for newly added functionality. The remaining
code is a refactor already covered by tests.

Change-Id: I9ad82c0757473e5a3311889cf49f73ff8857da02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385362
Commit-Queue: Andrew Davies <awdavies@google.com>
Testability-Review: Andrew Davies <awdavies@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>