[e2e] Creates a no-basemgr test
The "Screen is not black" test is intended to be ran on a device
configuration that supports taking screenshots, so that we can establish
whether there is something shown on screen when the time comes.

Some system configuration had basemgr running at boot, and some
configurations wer able to run basemgr, but didn't have it running.  The
test has code that tries to figure out whether basemgr is running and
tries to start it if it isn't.

This approach does not work on system configurations where basemgr is
not supposed to be running in the first place.  So, in this change, we
add a separate "no-basemgr" test, which assumes that the device has
another way of obtaining its screenshot and simply skips trying to
fiddle with basemgr.

This is not the most general approach, but based on what has been
reported in bug 50705, could be enough to make the test run well
in configurations that are currently of interest.

A side change, now that I am here, added canonical targets for the
directory and tests.

Bug: 50705
Change-Id: Ib06021078f9d99f768d2a4ebaba1fd877d882a22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386238
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: mesch <mesch@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Isma Cortes <isma@google.com>
Testability-Review: Filip Filmar <fmil@google.com>