[intl] Adds a build template for generating l10n
Adds a template `//build/intl/strings.gni:intl_strings` which takes on
most of the responsibility for generating localized resources from
localized material.

Since the template uses FIDL under the hood to generate polyglot
includable code, and given that this seems to be the first such use of
FIDL, I had to tweak the FIDL templates to allow `non_fidl_deps` to be
declared.

Since this all is a bit cumbersome still, I added an example to show how
one can deploy a Fuchsia package which compiles with the generated code
and bundles the localized resources in the Fuchsia package.  The
example will evolve with the sophistication of our L10N code.  The
example could probably use some trimming down, but at the moment I am
not sure about the approach.  So I left that for some followup work.

Fixed: 51172
Change-Id: I568c3252d35bc0d61fafadf3032b9a491bff796e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386391
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>