[fuchsia.hardware.input] Add DeviceReportReader
Add a new interface that reads reports. The
new interface allows the driver to be in control
of the new interface and have a HangingGet
function. It will also allow the instance drivers
to be removed in a later commit.

Take this time to have a Report struct,
so that a vector of Reports can be returned.

Bug: 49329
Test: hid-test

Change-Id: I75cfabc373b39d51020edb6ad013e5fc880921da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386295
Commit-Queue: David Gilhooley <dgilhooley@google.com>
API-Review: Ruchira Ravoori <ravoorir@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>